+++ src/IconBadges.hpp
@@ -4,7 +4,6 @@
 
 class IconBadges {
 public:
-    inline static std::map<int, std::map<IconType, std::vector<int>>> badges;
-
-    static bool enabled();
+    static std::map<int, std::map<IconType, std::vector<int>>> badges;
+    static bool enabled;
 };