+++ src/classes/IBBadgePopup.cpp
@@ -1,12 +1,13 @@
 #include "IBBadgePopup.hpp"
 #include "IBIconPopup.hpp"
+#include "../IconBadges.hpp"
 #include <Geode/binding/GJUserScore.hpp>
 
 using namespace geode::prelude;
 
-IBBadgePopup* IBBadgePopup::create(GJUserScore* score, const std::map<IconType, std::vector<int>>& icons) {
+IBBadgePopup* IBBadgePopup::create(GJUserScore* score) {
     auto ret = new IBBadgePopup();
-    if (ret->initAnchored(300.0f, 70.0f, score, icons)) {
+    if (ret->initAnchored(300.0f, 70.0f, score)) {
         ret->autorelease();
         return ret;
     }
@@ -14,7 +15,7 @@ IBBadgePopup* IBBadgePopup::create(GJUserScore* score, const std::map<IconType, 
     return nullptr;
 }
 
-bool IBBadgePopup::setup(GJUserScore* score, const std::map<IconType, std::vector<int>>& icons) {
+bool IBBadgePopup::setup(GJUserScore* score) {
     setID("IBBadgePopup");
     setTitle(fmt::format("{}'s Badges", GEODE_ANDROID(std::string)(score->m_userName)), "goldFont.fnt", 0.7f, 15.0f);
     m_title->setID("badges-title");
@@ -32,12 +33,13 @@ bool IBBadgePopup::setup(GJUserScore* score, const std::map<IconType, std::vecto
     badgeMenu->setLayout(RowLayout::create()->setGap(10.0f));
     badgeMenu->setID("badge-menu");
 
-    for (auto& [type, ids] : icons) {
+    for (auto& [type, ids] : IconBadges::badges[score->m_accountID]) {
+        if (ids.empty()) continue;
         auto badgeName = badgeNames[(int)type];
         auto badgeButton = CCMenuItemExt::createSpriteExtraWithFrameName(fmt::format("gj_{}Btn_on_001.png", badgeName), 0.7f, [
-            scoreRef = WeakRef(score), type, &ids
+            scoreRef = WeakRef(score), type
         ](auto) {
-            if (auto score = scoreRef.lock()) IBIconPopup::create(score, type, ids)->show();
+            if (auto score = scoreRef.lock()) IBIconPopup::create(score, type)->show();
         });
         badgeButton->setID(fmt::format("{}-badge-button", badgeName));
         badgeMenu->addChild(badgeButton);