+++ changelog.md
@@ -1,4 +1,8 @@
 # Icon Badges Changelog
+## v1.0.2 (2025-10-22)
+- Fixed a bug where robots and spiders in the icons popup would be mispositioned on switching pages
+- Fixed a bug where the badge popup would show badges for types of icons the user did not have
+
 ## v1.0.1 (2025-09-18)
 - Fixed badge menu position on leaderboard menu (Reported by [RayDeeUx](user:1941705))